Add options
One 15 second measurement does not measure previous data if the data only updates every 30 seconds. This commit gives the option to increase this value on systems which have slow update times for calibration purposes. It adds a before and during measurement time of 'dead_time' to the function one_measurement via a global variable.

When one_measurement is called, this change causes an effective delay before starting measurements so that the system changes recently made will be measured. It also adds time to the measurement process. This has no effect when a workload is run as this would require changing the workload itself. The main anticipated use of this function is to improve calibration.

Add command line parameter option of dead time.
Added global variable unsigned int dead_time:
This is similar to the time out option which is not used during calibration but has a similar purpose (not sure of exactly how it is used) - (perhaps merge the two later)

Added code to main to retrieve the command line option
Added code to one_measurement to add a before and during measurement delay of dead_time in seconds.

If this has unintended consequences the global variable can be passed to the calibration function. However since it may have utility outside of this function I have left it in for testing purposes.